Refactor mock_detector
